==> Mod3/jupyter_notebooks/01-01-numpy.ipynb <==
#  Instalar NumPy
📌Antes de ponernos a trabajar con NumPy tendremos que instalar la librería, para eso lo que tendremos que hacer es escribir en la terminal la siguiente línea de código.
```bash
!pip install numpy
```

# Creación de *arrays*

NumPy permite crear *arrays* de distintas formas las más comunes son:

1. **Crear un array a partir de una lista**: Puedes crear un *array* NumPy a partir de una lista utilizando la función `np.array()`.

2. **Crear un array vacío**: Puedes crear un *array* vacío especificando la forma deseada utilizando la función `np.empty()`. Debemos tener en cuenta que los valores del array vacío pueden ser impredecibles, ya que toman los valores almacenados en la memoria en ese momento.

3. **Crear un array de ceros**: Puedes crear un *array* con todos los elementos establecidos en cero utilizando la función `np.zeros()`. 

4. **Crear un array de unos**: Puedes crear un *array* con todos los elementos establecidos en uno utilizando la función `np.ones()`.

5. **Crear un array con valores secuenciales**: Puedes crear un *array* con valores secuenciales utilizando la función `np.arange()`. 

6. **Crear un array usando random de NumPy**: Puedes crear un *array* con números aleatorios usando distintos métodos de `random` en NumPy


# antes de empezar importamos la librería de NumPy para poder trabajar con todos sus métodos. 
import numpy as np

## Crear un array a partir de una lista

print("-------------ARRAY UNIDIMENSIONAL--------------")

# definimos una lista que usaremos para convertir en array unidimensional

lista = [1, 2, 3, 4, 5]
print("La lista a partir de la que crearemos un array unidimensional es:", lista)

# convertimos la lista en array usando el método 'np.array()'

array_unidimensional = np.array(lista)
print(f"El array unidimensional creado a partir de la lista es: \n  {array_unidimensional} \n")

print("\n -*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-\n ")

print("-------------ARRAY BIDIMENSIONAL--------------")
# para crear un array bidimensional necesitaremos una lista de listas

lista_listas = [[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]]
print("La lista de listas a partir de la que crearemos un array bidimensional es:", lista_listas)

# convertimos la lista en array usando el método 'np.array()' 
# 📌 fijaos como ahora el array que hemos construido tiene "filas" y "columnas"

array_bidimensional = np.array(lista_listas)
print(f"El array bidimensional creado a partir de la lista es:\n  {array_bidimensional}\n")

print("\n -*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-\n ")

print("-------------ARRAY TRIDIMENSIONAL--------------")
# para crear un array tridimensional necesitaremos una lista de listas de listas 🤯. 

lista_listas_listas = [[[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]], [[11, 12, 13, 14, 15], [16, 17, 18, 19, 20]]]
print("La lista de listas de listas a partir de la que crearemos un array tridimensional es:", lista_listas_listas)

# convertimos la lista en array usando el método 'np.array()' 
# fijaos como ahora el array que hemos construido tiene "filas" y "columnas"

array_tridimensional = np.array(lista_listas_listas)
print(f"El array tridimensional creado a partir de la lista es:\n  {array_tridimensional}\n")

Acabamos de aprender a crear *arrays* de distintas dimensiones, y en el print anterior vemos que tienen distintas formas, pero solo podemos decir eso por ahora. Exploremos con una serie de métodos específicos, cuales son las propiedades de cada uno de estos *arrays* para entender mejor sus diferencias. 

NumPy ofrece una variedad de métodos útiles para comprender y explorar las propiedades de un *array*. Veamos los más importantes:

1. **Forma del array**: El método `shape` devuelve una tupla que representa la forma (dimensiones) del *array*. Cada elemento de la tupla representa el tamaño de una dimensión. 

2. **Número de dimensiones**: El atributo `ndim` devuelve el número de dimensiones del *array*.

3. **Tamaño total del array**: El atributo `size` devuelve el número total de elementos en el *array*. 


4. **Tipo de datos del array**: El atributo `dtype` devuelve el tipo de datos de los elementos del *array*. 


print("-------------ARRAY UNIDIMENSIONAL--------------")
print(f"El array unidimensional es: \n  {array_unidimensional}")
print(f"La forma del array unidimensional es: {array_unidimensional.shape}")
print(f"El número de dimensiones del array unidimensional es: {array_unidimensional.ndim}")
print(f"El número de elementos del array unidimensional es: {array_unidimensional.size}")
print(f"El tipo de dato que tenemos en el array unidimensional es: {array_unidimensional.dtype}\n")

Entendamos los resultados que hemos obtenido en la celda anterior: 

- ***array* unidimensional**

    - *Forma*: el método `shape` nos devuelve una tupla donde nos va indicar el número de elementos que tenemos por cada dimensión que tiene nuestro *array*. En este caso, nos devuelve una tupla de un solo elemento porque solo tenemos una dimensión que tiene 5 elementos. 

    - *Número de dimensiones*: para entender bien que es lo que nos devuelve el atributo `ndim` necesitamos comprender que es una dimensión en el contexto de un *array*. En términos simples, una dimensión en un *array* se refiere a la cantidad de índices necesarios para acceder a un elemento dentro del *array*. Por ejemplo, un *array* unidimensional requiere un solo índice para acceder a un elemento, mientras que un *array* bidimensional requiere dos índices (uno para la fila y otro para la columna). Entendiendo esto,  y pensando que la indexación en *arrays* es como en las listas, en nuestro caso solo necesitamos un índice para acceder a algún valor dentro del *array*, por lo que la dimensión de este *array* es de 1

    - *Número de elementos*: con el atributo `size` nos dice el número de elementos que  hay dentro del *array*, en este caso tenemos 5 números enteros

    - *Tipo de datos*: usando el atributo `dtype` conocemos que  el *array* esta compuesto por elementos de tipo *integer*. Recordad que en un *array* todos los elementos deben ser del mismo tipo. 


print("-------------ARRAY BIDIMENSIONAL--------------")
print(f"El array bidimensional es: \n  {array_bidimensional}")
print(f"La forma del array bidimensional es: {array_bidimensional.shape}")
print(f"El número de dimensiones del array bidimensional es: {array_bidimensional.ndim}")
print(f"El número de elementos del array bidimensional es: {array_bidimensional.size}")
print(f"El tipo de dato que tenemos en el array bidimensional es: {array_bidimensional.dtype}\n")

Entendamos los resultados que hemos obtenido en la celda anterior: 


- ***array* bidimensional**

    - *Forma*: en este caso la tupla esta compuesta por dos valores, podríamos decir que tenemos "filas" y "columnas". En concreto nos esta diciendo que tenemos 2 filas y 5 columnas. 

    - *Número de dimensiones*: En este caso necesitamos dos índices para acceder a algún valor dentro del *array*, por lo que la dimensión de este *array* es de 2. Es como acceder a un elemento en una lista de listas

    - *Número de elementos*: con el atributo `size` nos dice el número de elementos que  hay dentro del *array*, en este caso tenemos 10 números enteros

    - *Tipo de datos*: usando el atributo `dtype` conocemos que  el *array* esta compuesto por elementos de tipo *integer*. Recordad que en un *array* todos los elementos deben ser del mismo tipo. 


print("-------------ARRAY TRIDIMENSIONAL--------------")
print(f"El array tridimensional es: \n  {array_tridimensional}")
print(f"La forma del array tridimensional es: {array_tridimensional.shape}")
print(f"El número de dimensiones del array tridimensional es: {array_tridimensional.ndim}")
print(f"El número de elementos del array tridimensional es: {array_tridimensional.size}")
print(f"El tipo de dato que tenemos en el array tridimensional es: {array_tridimensional.dtype}\n")

Entendamos los resultados que hemos obtenido en la celda anterior: 

- ***array* tridimensional**

    - *Forma*: en este caso la tupla esta compuesta por tres valores. Esto lo podemos interpretar de la siguiente forma, tenemos 3 dimensiones, la primera hace referencia al número de *arrays* más pequeños que tenemos en nuestro *array* principal (en este caso 2, es decir, tenemos dos *arrays*). Luego tenemos dos valores más que son el 2 y el 5 que indican que cada uno de los *arrays* pequeños tienen dos filas y 5 columnas. 

    - *Número de dimensiones*: En este caso necesitamos tres índices para acceder a algún valor dentro del *array*, por lo que la dimensión de este *array* es de 3. 

    - *Número de elementos*: con el atributo `size` nos dice el número de elementos que  hay dentro del *array*, en este caso tenemos 20 números enteros

    - *Tipo de datos*: usando el atributo `dtype` conocemos que  el *array* esta compuesto por elementos de tipo *integer*. Recordad que en un *array* todos los elementos deben ser del mismo tipo. 


Comprendiendo las diferencias entre los distintos tipos de *arrays* y sus propiedades, sigamos viendo otras formas de crear **arrays**. 

## Crear un array vacío

Al igual que en el caso anterior, podremos crear *arrays* vacíos de distintas dimensiones. Su sintaxis básica es:


```python
np.empty(shape, dtype=float)
```
- `shape`: Especifica la forma (dimensiones) del *array* que se creará. Puede ser un entero, una tupla de enteros o una lista de enteros. 

- `dtype` (opcional): Especifica el tipo de datos de los elementos del *array*. Por defecto, es `float`, lo que significa que los elementos del *array* serán números de punto flotante. También se pueden especificar otros tipos de datos, como `int` para números enteros, `bool` para valores booleanos, `str` para cadenas de texto, entre otros.


Aquí tienes algunos ejemplos de cómo utilizar `np.empty`:

```python
import numpy as np

# Crear un array vacío de 2 filas y 3 columnas
array_vacio = np.empty((2, 3))
print(array_vacio)

# Crear un array vacío de forma (4, )
array_vacio_1D = np.empty(4)
print(array_vacio_1D)
```

El método `np.empty` crea un *array* vacío sin inicializar sus elementos. Esto significa que los valores de los elementos pueden ser impredecibles y dependerán de los datos almacenados previamente en la memoria.

print("------------- ARRAY UNIDIMENSIONAL VACÍO --------------")

# creamos el array vacío unidimensional especificando el número de elementos que queremos tener en el array. 
# como solo tiene una dimensión solo le paso un valor. Nos genera un array con 4 valores.

array_vacio_unidimensional =  np.empty(4)
print(f"El array unidimensional creado a partir del método np.emtpy(): \n  {array_vacio_unidimensional} \n")

print("------------- ARRAY BIDIMENSIONAL VACÍO --------------")

# creamos el array vacío bidimensional especificando el número de elementos que queremos tener en el array. 
# como en este caso queremos tener dos dimensiones, le tendremos que pasar dos valores, con el número de elementos que queremos en cada una de ellas
# el primer valor hace referencia al número de filas y el segundo al número de columnas.

array_vacio_bidimensional =  np.empty((4,2))
print(f"El array bidimensional creado a partir del método np.emtpy(): \n  {array_vacio_bidimensional} \n")


print("------------- ARRAY TRIDIMENSIONAL VACÍO --------------")

# creamos el array vacío tridimensional especificando el número de elementos que queremos tener en el array. 
# queremos tener tres dimensiones, por lo que le tendremos que pasar tres valores, con el número de elementos que queremos en cada una de ellas
# el primer valor hace referencia al número de arrays "pequeños" que queremos, el segundo al número de filas y el último al número de columnas

array_vacio_tridimensional =  np.empty((3, 4, 2)) # estamos generando un array de tres matrices donde cada una de ellas tienen 4 filas y 2 columnas.
print(f"El array tridimensional creado a partir del método np.emtpy(): \n  {array_vacio_tridimensional} \n")

## Crear un array de ceros

Permite crear un *array* compuesto solo por 0. La sintaxis del método `np.zeros` en NumPy es la siguiente:

```python
np.zeros(shape, dtype=float)
```

- `shape`: Especifica la forma (dimensiones) del *array* que se creará. Puede ser un entero, una tupla de enteros o una lista de enteros.

- `dtype` (opcional): Especifica el tipo de datos de los elementos del *array*. Por defecto, es `float`, lo que significa que los elementos del *array* serán números de punto flotante. También se pueden especificar otros tipos de datos, como `int` para números enteros, `bool` para valores booleanos, `str` para cadenas de texto, entre otros.



print("------------- ARRAY UNIDIMENSIONAL DE CEROS --------------")

# creamos el array unidimensional especificando el número de elementos que queremos tener en el array. 
# como solo tiene una dimensión solo le paso un valor.

array_vacio_unidimensional_cero =  np.zeros(4)
print(f"El array unidimensional creado a partir del método np.zeros(): \n  {array_vacio_unidimensional_cero} \n")

print("------------- ARRAY BIDIMENSIONAL DE CEROS --------------")

# la lógica y la sintaxis va a ser la misma que cuando creabamos los arrays vacíos
# creamos un array de 0 con 4 filas y 2 columnas.

array_vacio_bidimensional_cero =  np.zeros((4,2))
print(f"El array bidimensional creado a partir del método np.zeros(): \n  {array_vacio_bidimensional_cero} \n")

print("------------- ARRAY TRIDIMENSIONAL DE CEROS --------------")

# estamos generando un array de 0 con 3 matrices donde cada una de ellas tienen 4 filas y 2 columnas.

array_vacio_tridimensional_cero =  np.zeros((3, 4, 2)) 
print(f"El array tridimensional creado a partir del método np.zeros(): \n  {array_vacio_tridimensional_cero} \n")


## Crear un array de unos

Crea un *array* formado unicamente por 1. La sintaxis del método `np.ones` en NumPy es la siguiente:

```python
np.ones(shape, dtype=None)
```

- `shape`: Especifica la forma (dimensiones) del array que se creará. Puede ser un entero, una tupla de enteros o una lista de enteros. 


- `dtype` (opcional): Especifica el tipo de datos de los elementos del array. Por defecto, es `None`, lo que significa que los elementos del array serán números de punto flotante. También se pueden especificar otros tipos de datos, como `int` para números enteros, `bool` para valores booleanos, `str` para cadenas de texto, entre otros.

print("------------- ARRAY UNIDIMENSIONAL DE UNOS --------------")

# creamos el array unidimensional especificando el número de elementos que queremos tener en el array. 
# como solo tiene una dimensión solo le paso un valor.

array_vacio_unidimensional_unos =  np.ones(4)
print(f"El array unidimensional creado a partir del método np.ones(): \n  {array_vacio_unidimensional_unos} \n")

print("------------- ARRAY BIDIMENSIONAL DE UNOS --------------")

# estamos creando un array de dos dimensiones con 4 filas y 2 columnas

array_vacio_bidimensional_unos =  np.ones((4,2))
print(f"El array bidimensional creado a partir del método np.ones(): \n  {array_vacio_bidimensional_unos} \n")

print("------------- ARRAY TRIDIMENSIONAL DE UNOS --------------")

# creamos el array tridimensional especificando el número de elementos que queremos tener en el array. 

array_vacio_tridimensional_unos =  np.ones((3, 4, 2)) # estamos generando un array de tres matrices donde cada una de ellas tienen 4 filas y 2 columnas
print(f"El array tridimensional creado a partir del método np.ones(): \n  {array_vacio_tridimensional_unos} \n")

## Crear un array con valores secuenciales (`np.arange`)


En Numpy tenemos el metodo `np.arange` que funciona de forma similar al `range()` de Python, pero en este caso nos genera un *array*. La sintaxis del método `np.arange` es la siguiente:

```python
np.arange(start, stop, step, dtype=None)
```

- `start` (opcional): Especifica el valor de inicio de la secuencia. Por defecto, es 0.

- `stop`: Especifica el valor final de la secuencia (exclusivo). Es decir, la secuencia generada no incluirá este valor.

- `step` (opcional): Especifica el incremento (salto) entre los valores de la secuencia. Por defecto, es 1.

- `dtype` (opcional): Especifica el tipo de datos de los elementos en el array. Por defecto, es `None`, lo que significa que se utilizará el tipo de datos predeterminado.


📌 Por defecto nos devuelve una matriz unidimensional. En caso de que queramos cambiar las dimensiones del *array* tendremos que aplicar el método `reshape` que veremos mas adelante. 

print("------------- ARRAY UNIDIMENSIONAL CON 4 ELEMENTOS --------------")

array_unidimensional_arange_1 =  np.arange(4)
print(f"El array unidimensional creado a partir del método np.emtpy(): \n  {array_unidimensional_arange_1} \n")


print("------------- ARRAY UNIDIMENSIONAL CON NÚMEROS DEL 1 AL 4 (no incluido) CON SALTOS DE 1 EN 1  --------------")
array_unidimensional_arange_2 =  np.arange(1,4,1)
print(f"El array bidimensional creado a partir del método np.emtpy(): \n  {array_unidimensional_arange_2} \n")


print("------------- ARRAY UNIDIMENSIONAL CON NÚMEROS DEL 1 AL 20 (no incluido) CON SALTOS DE 2 EN 2  --------------")
array_unidimensional_arange_3 =  np.arange(1,20,2)
print(f"El array bidimensional creado a partir del método np.emtpy(): \n  {array_unidimensional_arange_3} \n")


## Crear un array usando random

El módulo `random` de NumPy proporciona funciones para generar números aleatorios con diferentes distribuciones y características. Estas funciones son útiles para simular datos aleatorios y realizar experimentos numéricos. Algunos de los métodos más utilizados del módulo `random` de NumPy:


1. `np.random.randint`: Genera números enteros aleatorios en un array. Puedes especificar el rango de los números enteros y la forma del array como argumentos.

2. `np.random.rand`: Genera números aleatorios en un *array* en el rango [0, 1).  Es decir, los números generados estarán en el intervalo semiabierto [0, 1), donde 0 está incluido y 1 está excluido.

3. `np.random.sample`: Genera números aleatorios en un *array*  en el rango [0, 1).


### np.random.randint

La sintaxis básica del método `np.random.randint` es la siguiente:

```python
numpy.random.randint(low, high, size, dtype)
```

Donde:

- `low`: Especifica el límite inferior del rango de números enteros a generar (inclusive). Si solo se proporciona este argumento, los números aleatorios se generarán en el rango [0, low). Es decir, el límite superior no está incluido en el rango.

- `high` (opcional): Especifica el límite superior del rango de números enteros a generar (exclusivo). Si se proporciona este argumento, los números aleatorios se generarán en el rango [low, high). 

- `size` (opcional): Especifica la forma del *array* resultante de números enteros aleatorios. Puede ser un número entero o una tupla de enteros que represente las dimensiones del *array*. Por defecto, es `None`, lo que implica que se generará un solo número entero,  es decir, un array unidimensional.

- `dtype` (opcional): Especifica el tipo de dato de los elementos en el *array*. Por defecto, es `int`, lo que significa que los números enteros se generarán como enteros estándar.



print("------------- ARRAY UNIDIMENSIONAL RANDINT --------------")
# en este caso estamos generando un array unidimensional con un valor aleatorio que puede variar entre 0 y 9 (no incluido)

array_randint_uni = np.random.randint(9)
print(f"El array unidimensional creado a partir del método np.random.randint(): \n  {array_randint_uni} \n")

print("------------- ARRAY BIDIMENSIONAL RANDINT --------------")
# generamos un array bidimensional con números aleatorios entre el 0 y el 50 (no incluido) de 2 filas y 3 columnas

array_randint_bi = np.random.randint(0, 50, (2,3))
print(f"El array bidimensional creado a partir del método np.random.randint(): \n  {array_randint_bi} \n")

print("------------- ARRAY TRIDIMENSIONAL RANDINT --------------")
# generamos un array tridimensional de números aleatorios entre el 0 y 20 (no incluido) de 2 matrices, 3 filas y 4 columnas

array_randint_tri = np.random.randint(0, 20, (2, 3, 4))
print(f"El array bidimensional creado a partir del método np.random.randint(): \n  {array_randint_tri} \n")

### np.random.rand

La sintaxis básica del método `np.random.rand` es la siguiente:

```python
numpy.random.rand(d0, d1, ..., dn)
```

- `d0, d1, ..., dn` son los argumentos opcionales que determinan las dimensiones de la matriz de números aleatorios que se generará. Puedes proporcionar cualquier número de argumentos, lo que significa que puedes crear matrices unidimensionales, bidimensionales, tridimensionales, etc.

📝 **NOTA**: en este caso no tenemos que especificar que entre que valores queremos que nos genere los números. POR DEFECTO NOS GENERARÁ NÚMEROS ALEATORIOS ENTRE EL 0-1.

📝 **NOTA**: Solo nos dará *floats*, por lo que no hace falta especificar el tipo. 

print("------------- ARRAY UNIDIMENSIONAL RAND --------------")
# en este caso estamos generando un array unidimensional con cuatro valores aleatorios entre el 0 y el 1

array_rand_uni = np.random.rand(4)
print(f"El array unidimensional creado a partir del método np.random.rand(): \n  {array_rand_uni} \n")

print("------------- ARRAY BIDIMENSIONAL RAND --------------")
# en este caso estamos generando un array bidimensional de 2 filas y 3 columnas con números aleatorios entre el 0 y el 1

array_rand_bi = np.random.rand(2,3)
print(f"El array unidimensional creado a paratir del método np.random.rand(): \n  {array_rand_bi} \n")

print("------------- ARRAY TRIDIMENSIONAL RAND --------------")
# en este caso estamos generando un array tridimensional de 2 matrices, 3 filas  y 4 columnas con números aleatorios entre el 0 y el 1

array_rand_tri = np.random.rand(2,3,4)
print(f"El array unidimensional creado a partir del método np.random.rand(): \n  {array_rand_tri} \n")

### np.random.random_sample

La sintaxis básica del método `np.random.random_sample` es la siguiente:

```python
numpy.random.random_sample(size=None)
```

Donde:

- `size` (opcional): Permite especificar la forma del array resultante. Puede ser un número entero o una tupla de enteros que represente las dimensiones del array. Si no se proporciona el parámetro `size`, se devuelve un solo número aleatorio en el rango [0, 1).

print("------------- ARRAY UNIDIMENSIONAL RANDOM_SAMPLE --------------")
# en este caso estamos generando un array unidimensional con 4 valores aleatorios entre el 0 y el 1

array_sample_uni = np.random.random_sample(4)
print(f"El array unidimensional creado a partir del método np.random.random_sample(): \n  {array_sample_uni} \n")

print("------------- ARRAY BIDIMENSIONAL RANDOM_SAMPLE --------------")
# en este caso estamos generando un array bidimensional de 2 filas y 3 columnas con números aleatorios entre el 0 y el 1

array_sample_bi = np.random.random_sample((2,3))
print(f"El array unidimensional creado a paratir del método np.random.random_sample(): \n  {array_sample_bi} \n")

print("------------- ARRAY TRIDIMENSIONAL RANDOM_SAMPLE --------------")
# en este caso estamos generando un array tridimensional de 2 matrices, 3 filas  y 4 columnas con números aleatorios entre el 0 y el 1

array_sample_tri = np.random.random_sample((2,3, 4))
print(f"El array unidimensional creado a partir del método  np.random.random_sample(): \n  {array_sample_tri} \n")
