==> Mod3/jupyter_notebooks/02-01-series.ipynb <==
# importamos la pandas para poder trabajar en la lección
import pandas as pd

**Serie vacía**

# Creemos algunas Series en Pandas para entenderlas mejor.
# en este primer ejemplo crearemos una Serie vacía
serie_vacia = pd.Series()
print("mi serie de pandas es:",serie_vacia)
print("los tipos de elementos que la componen son:",serie_vacia.dtype)

**Serie a partir de listas**

# esta Serie esta bien, pero en realidad por lo general queremos tener Series con datos
# creemos ahora una serie a partir de una lista

# definimos la lista
lista = [23, 45, 17, 83, 67]

# convertimos esta lista en una Serie
serie_lista1 = pd.Series(lista)
serie_lista1

📌 Si nos fijamos el método `pd.Series` nos va a generar automáticamente el índice y lo crea con números consecutivos, empezando por el 0. Sin embargo, también le podemos pasar los índices que queramos usando el parámetro `index`. 

> index: Es un parámetro opcional de la función `pd.Series`. En este caso, se proporciona una lista de índices personalizados para la Serie. Los índices indican cómo se etiquetarán los elementos de la Serie. En este ejemplo, los índices se definen en el orden especificado: 3, 4, 5, 6, 7.

Tener índices personalizados puede ser útil cuando necesitamos acceder a los elementos de la Serie utilizando etiquetas específicas en lugar de las posiciones numéricas predeterminadas. Esto facilita la manipulación y el análisis de los datos en la Serie. 

Es importante destacar que los índices pueden ser de diferentes tipos, incluidos *integer*, cadenas, fechas, etc. Además, los índices no necesariamente tienen que estar en orden o ser consecutivos. Esto proporciona una gran flexibilidad en el uso de las Series en Pandas.

# vamos a definir una nueva serie pero en este caso vamos a ponerle nuestro propio índice usando el parámetro 'index'
serie_lista2 = pd.Series(lista, index = [3,4,5,6,7])
serie_lista2

# como hemos mencionado los índices pueden ser de distintos tipos, en este caso crearemos una nueva Serie donde sus índices sean strings
serie_lista3 = pd.Series(lista, index = ["Lunes","Martes","Miercoles","Jueves","Viernes"])
serie_lista3

**Series a partir de diccionarios**

# definir un diccionario 
dicc = {'lorena' : 10, 
        'marta' : 20, 
        'pilar' : 30, 
        'laura': 50, 
       'ana': 86, 
        'maria': 28} 

# crear una serie nueva a partir del diccionario definido previamente
serie_diccionario = pd.Series(dicc)
serie_diccionario

## Propiedades de las Series

Las propiedades son atributos específicos que nos proporcionan información sobre las características de una Serie en particular. Las propiedades más importantes de las Series en Pandas son:

1. **values**: Esta propiedad devuelve los valores de la Serie como un array NumPy. Puedes acceder a ellos utilizando la sintaxis `serie.values`.

2. **index**: Esta propiedad devuelve los índices de la Serie. Los índices pueden ser de cualquier tipo, incluyendo *integer*s, cadenas u otros objetos. Puedes acceder a ellos utilizando la sintaxis `serie.index`.

3. **dtype**: Esta propiedad devuelve el tipo de datos de los elementos en la Serie. Puedes acceder a él utilizando la sintaxis `serie.dtype`. Los tipos de datos comunes incluyen `int64`, `float64`, `object`, `datetime64`, entre otros.

4. **size**: Esta propiedad devuelve el número total de elementos en la Serie. Puedes acceder a él utilizando la sintaxis `serie.size`.

5. **shape**: Esta propiedad devuelve una tupla que representa la forma (dimensiones) de la Serie. En el caso de una Serie unidimensional, la forma será `(n,)`, donde `n` es el número de elementos en la Serie. Puedes acceder a él utilizando la sintaxis `serie.shape`.

Estas propiedades son útiles para obtener información sobre una Serie en particular, como el tipo de datos, el número de elementos y los índices asociados. Además, las propiedades pueden ser utilizadas en conjunto con los métodos y funciones de Pandas para realizar operaciones y análisis más avanzados en los datos de la Serie.

Es importante tener en cuenta que las Series son inmutables, lo que significa que no puedes cambiar sus elementos directamente. Sin embargo, puedes realizar operaciones y manipulaciones en las Series para obtener nuevos resultados o modificar la estructura de los datos en función de tus necesidades.

# recordemos la serie
display("La serie con la que vamos a trabajar es:", serie_diccionario)
print("---------------------------")

# utilizamos el método index que nos va a devolver todos los índices 
# fijaos como en este caso nos ha devuelto todos los índices de nuestra Serie, es decir, los nombres de las alumnas
print(f"El método 'index' nos devuelve los índices de la Serie: {serie_diccionario.index}")
print("---------------------------")

# ahora utilizamos el método 'values' nos devuelve todos los valores de la Serie
# en este caso los números que teníamos
print(f"El método 'values' nos devuelve los valores de la Serie: {serie_diccionario.values}")
print("---------------------------")

#utilizando el método 'shape' nos devuelve el número de "filas" que tenemos en la Serie
print(f"El método 'shape' nos devuelve la forma de la Serie: {serie_diccionario.shape}")
print("---------------------------")

#utilizando el método 'size' nos devuelve el número de elementos que tenemos en la Serie
print(f"El método 'size' nos devuelve el número de elementos de la Serie: {serie_diccionario.size}")
print("---------------------------")

#utilizando el método 'dtypes' nos devuelve el tipo de datos de la Serie
print(f"El método 'dtypes' nos devuelve el tipo de datos de la Serie: {serie_diccionario.dtypes}")



## Indexación de las Series

La indexación se refiere a la forma en que accedemos a los elementos de una Serie utilizando índices o etiquetas. Pandas ofrece varias opciones para indexar y seleccionar elementos en una Serie. Algunas de las principales formas de realizar la indexación en las Series son:

1. **Indexación por posición**: Podemos acceder a los elementos de una Serie utilizando la posición numérica del elemento en la Serie, como hacíamos con las listas

2. **Indexación por etiqueta**: Podemos acceder a los elementos de una Serie utilizando las etiquetas de índice asociadas a cada elemento. 

3. **Indexación por rango**: Podemos acceder a un rango de elementos de una Serie utilizando los índices de inicio y fin. 

4. **Indexación por lista de índices**: Podemos acceder a varios elementos de una Serie utilizando una lista de índices.


 **Indexación por posición**

display("La serie con la que vamos a trabajar es:", serie_diccionario)
print("---------------------------")

# accedemos al primer elemento (el valor 10). 🚨 Fijaos como los índices son como los de las listas, empiezan en 0
print("El valor que está en posición 0 es:", serie_diccionario[0])  
print("---------------------------")

# accedemos al tercer elemento (el valor 30)
print("El valor que está en posición 2 es:", serie_diccionario[2])  


**Indexación por etiqueta**

display("La serie con la que vamos a trabajar es:", serie_diccionario)
print("---------------------------")


# accedemos al valor de la etiqueta de "laura" (el valor 50). 🚨 Esto lo podríamos comparar con los diccionarios
print("El valor que está en indice 'laura' es:", serie_diccionario["laura"])  
print("---------------------------")

# accedemos al valor de la etiqueta de "pilar" (el valor 30). 
print("El valor que está en indice 'pilar' es:", serie_diccionario["pilar"])  

**Indexación por rango**: 

display("La serie con la que vamos a trabajar es:", serie_diccionario)
print("---------------------------")

# accedemos a los elementos desde el índice 1 al 3 ([20, 30, 50]). 🚨 De niuevo fijaos como esto funciona igual que en listas
# nos esta incluyendo el valor que está en el índice 1, pero no el que está en el índice 4
print("Los valores de los índices que están entre las posiciones 1 y 3 (incluida) es: \n", serie_diccionario[1:4])  

**Indexación por lista de índices**

display("La serie con la que vamos a trabajar es:", serie_diccionario)
print("---------------------------")

# Acceder a los elementos con índices 1, 3 y 4. 🚨 Fijaos como en este caso, le hemos puesto dos corchetes
print("Los valores de los índices que están en las posiciones 1, 3 y 4 son: \n", serie_diccionario[[0,2,3]])
