==> Mod3/jupyter_notebooks/02-04-crear-columnas.ipynb <==
# importamos la pandas para poder trabajar en la lección
import pandas as pd

# create a sample data.csv file
sample_data = {
	"Day": ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday"],
	"Temperature": [20, 21, 19, 18, 22]
}
sample_df = pd.DataFrame(sample_data)
sample_df.to_csv("data.csv", index=False)

# importamos un nuevo fichero de datos para entender mejor estos conceptos
df = pd.read_csv("data.csv", index_col=["Day"])
df

### Asignación directa

Supongamos que queremos tener la temperatura también en grados Farenheit. Podríamos crearnos una columna nueva aplicando una serie de operaciones sobre la que ya tenemos: 

# lo primero que hacemos es establecer el nombre de la nueva columna que queremos, y esto lo igualaremos a la operación que queremos realizar.
# en este caso, todos los valores de la columna 'Temperature' multiplicados por 9/5 y a todo esto le sumamos 32
# lo que va a hacer es ir valor a valor (celda a celda) aplicando las operaciones que le hemos especificado. 

df["Farenheit"] = (df["Temperature"] * 9/5) + 32

# si ahora volvemos a ver nuestro dataframe tendremos una nueva columna
# con la temperatura convertida en Farenheit

df.head()

### `.assign()`

La sintaxis básica del método `assign()` de Pandas es la siguiente:

```python
df_new = df.assign(nueva_columna=expresión)
```

# creamos una nueva columna llamada `temp_f` que será el resultado de la operación que hemos explicado previamente
df.assign(temp_f=df['Temperature'] * 9 / 5 + 32)

# comprobamos que en el  DataFrame se ha creado.... 
df

# Pero ... ¿Qué ha pasado? la ha creado, pero no la ha guardado (ver el output de arriba donde me falta la columna creada con assign), para eso tendremos que sobreescribir el DataFrame
df = df.assign(temp_f=df['Temperature'] * 9 / 5 + 32)

# comprobamos de nuevo si la columna se ha creado
df

### `insert()`

La sintaxis básica del método `insert()` es la siguiente:

```python
df.insert(loc, column, value, allow_duplicates=False)
```

# en este ejemplo estamos creando una columna nueva que se llama "indice" en la primera columna con números del 1 al 7. 

df.insert(0, "indice", range(1,8))

# comprobamos si la coolumna se ha creado correctamente
df

# ⚠️ Si volvieramos a crear otra colummna con el nombre de una columna ya existente nos daría error, 
# porque no podemos insertar dos columnas con el mismo nombre, ya que el parámetro `allow_duplicates` esta puesto en False
# incluso si la crearamos en otra posicion, seguiriamos teniendo el error. 

df.insert(0, "indice", range(1,8))
