==> Mod3/jupyter_notebooks/03-01-EDA-entender-valores.ipynb <==
# El *dataset* que vamos a utilizar

Los datos están relacionados con las campañas de *marketing* de una institución bancaria portuguesa. Las campañas de marketing se basaron en llamadas telefónicas. A menudo, se requería más de un contacto con el mismo cliente para determinar si el producto (depósito a plazo bancario) sería suscrito o no. Las colummas que tenemos en este *dataset* son: 

- **age**: La edad del cliente.

- **job**: La ocupación o profesión del cliente.

- **marital**: El estado civil del cliente.

- **education**: El nivel educativo del cliente.

- **default**: Indica si el cliente tiene algún historial de incumplimiento de pagos (1: Sí, 0: No).

- **housing**: Indica si el cliente tiene un préstamo hipotecario (1: Sí, 0: No).

- **loan**: Indica si el cliente tiene algún otro tipo de préstamo (1: Sí, 0: No).

- **contact**: El método de contacto utilizado para comunicarse con el cliente.

- **duration**: La duración en segundos de la última interacción con el cliente.

- **campaign**: El número de contactos realizados durante esta campaña para este cliente.

- **emp.var.rate**: La tasa de variación del empleo.

- **cons.price.idx**: El índice de precios al consumidor.

- **cons.conf.idx**: El índice de confianza del consumidor.

- **euribor3m**: La tasa de interés de referencia a tres meses.

- **nr.employed**: El número de empleados.

- **y**: Indica si el cliente ha suscrito un producto o servicio (Sí/No).

- **date**: La fecha en la que se realizó la interacción con el cliente.

- **latitude**: La latitud de la ubicación asociada al cliente.

- **longitude**: La longitud de la ubicación asociada al cliente.

- **id_**: Un identificador único para cada registro en el dataset.



## Entender las variables/columnas del *dataset*

En este *dataset* tenemos las siguientes columnas:

# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd

# cargamos el csv con el que vamos a ir desarrollando la lección de hoy, recordad que la descripción la tenéis al inicio del jupyter. 
# para eso usaremos el método "pd.read_csv" que ya vimos en la leccion anterior.

df = pd.read_csv("files/bank-additional.csv", index_col = 0)

# hacemos un ".head()" para mostrar las 5 primeras filas del dataframe
df.head()

# oh oh... si nos fijamos bien, no podemos ver todas las columnas del DataFrame (hay '...' en mitad de nuestra tabla) 
# ¿qué podemos hacer para conseguir que se vean todas las gráficas? 

pd.set_option('display.max_columns', None)

# después de aplicar el método ya podremos ver todas las columnas del df.

df.head()

# recordemos que también tenemos el método ".tail()", el cual nos mostrará las últimas 5 filas del dataframe

df.tail()

# a veces puede que no nos interese mostrar las 5 primeras o 5 últimas filas de un dataframe, y puede que solo queramos mostrar las dos primeras o las dos últimas
# esto lo podemos hacer también con ".head()" o ".tail()", incluyendo el número de filas que queremos ver dentro del paréntesis

df.head(2) # en este caso mostramos solo las dos primeras filas

# o solo la últimas 3 filas

df.tail(3)

# ¿y si quisieramos ver "x" filas al azar de nuestro conjunto de datos? Podemos usar el método ".sample()", el cual por defecto nos devuelve solo una fila

df.sample()

# si quisieramos ver 4 filas al azar de nuestro conjunto de datos solo tendríamos que especificarlo dentro de los paréntesis

df.sample(4)

# Ahora presentamos otro metodo que nos servira para conocer mejor nuestro dataframe, el método".shape" 
# que nos permitía saber cuántas filas y columnas tenemos en el DataFrame. 
# 📌 Recordemos que nos devuelve una tupla! donde su primer valor son las filas y el segundo valor las columnas.

print(f"El número de filas que tenemos es {df.shape[0]}, y el número de columnas es {df.shape[1]}")

# otro de los métodos que más usaremos es el método que nos devuelve el nombre de todas las columnas de nuestro DataFrame. Este es el método '.columns' 
# nos va a devolver un array con todos los nombres de las columnas

df.columns

Puede suceder que tengamos un dataframe con una cantidad enorme de columnas, y nosotras queremos saber la posicion de una columna en concreto, suponagmos la "duration". Podemos usar el metodo columns.get_loc para conocer su posicion dentro de la lista de columnas.

# Conociendo la posicion de una columna en particular

df.columns.get_loc("duration") # el resultado sera 8 porque es su posicion.

Sigamos aprendiendo nuevos métodos de Pandas que nos pueden ayudar en este análisis exploratorio de los datos, el método `.describe()`. Es una herramienta esencial para obtener una visión general de las estadísticas descriptivas de un DataFrame, lo que facilita la comprensión de la distribución de datos. Este método proporciona una instantánea concisa de las principales características numéricas de las columnas, como la media, la desviación estándar, los valores mínimo y máximo, los percentiles y más.

#utilizamos el metodo '.describe()' para sacar los principales estadísticos del DataFrame
# por defecto, nos devolverá solo los estadísticos de las VARIABLES NUMÉRICAS

df.describe()

# podemos usar el método '.T' para transponer los resultados

df.describe().T

**¿Qué es lo que vemos aquí?**

Entendamos los resultados de algunas de las columnas, por ejemplo `age`: 

- `count`: El número de observaciones no nulas en la columna "age".

- `mean`: La media de las edades en la columna.

- `std`: La desviación estándar, que mide la dispersión de las edades.

- `min`: La edad mínima en el conjunto de datos.

- `25%`: El percentil 25, que representa el valor por debajo del cual se encuentra el 25% de las edades.

- `50%`: La mediana o percentil 50, que es el valor que divide el conjunto de datos en dos mitades iguales.

- `75%`: El percentil 75, que representa el valor por debajo del cual se encuentra el 75% de las edades.

- `max`: La edad máxima en el conjunto de datos.



# hagamos lo mismo para las columnas categóricas, para eso tendremos que incluir entre los paréntesis el parámetro include = "object"

df.describe(include = "object").T

**Entendamos los resultados para la columna `job`**


- `count`: El número de observaciones no nulas en la columna "job".

- `unique`: La cantidad de valores únicos en la columna "job".

- `top`: El valor más común en la columna "job" (en este caso, "admin.").

- `freq`: La frecuencia del valor más común (en este caso, "admin.").


En el análisis preliminar de datos, existen dos métodos muy interesantes: `unique()` y `.value_counts()`.

El método `unique()` se utiliza para obtener los valores únicos de una columna. Devuelve un *array* de los valores distintos presentes en esa columna, sin repetir. Será útil para identificar las categorías únicas en una variable categórica o los diferentes elementos en una columna de texto. 

Por otro lado, el método `.value_counts()` nos permite contar la frecuencia de cada valor en una columna. Devuelve una serie que muestra el valor como índice y el recuento como valor. Es útil para obtener información sobre la distribución de los valores en una columna, especialmente en variables categóricas. 

Ambos métodos son herramientas poderosas para explorar y comprender la naturaleza de los datos, y nos permiten obtener información clave sobre la distribución y la unicidad de los valores en nuestras variables.

# recordemos el DataFrame

df.head(1)

# saquemos los valores únicos para la columna de `marital` usando el método 'unique()'

df["marital"].unique()

# y si usamos el '.value_counts()' 
# en este caso vemos que dentro de nuestro conjunto de datos tenemos 25999 casados, 12105 solteros y 4811 divorciados

df["marital"].value_counts()

Genial!! Pero solo lo hemos hecho para una columna, en realidad nos interesaría hacerlo para todas las variables categóricas del DataFrame, ya que en realidad queremos es explorar todas las columnas categóricas. Por lo tanto, lo que tendremos que hacer es: 

- Seleccionar las variables que son categóricas de nuestro DataFrame, para esto usaremos el método `.select_dtypes()` que es una función en Pandas que nos permite seleccionar columnas de un DataFrame basándose en su tipo de datos. Devuelve un DataFrame que contiene solo las columnas que cumplen con los tipos de datos especificados en los argumentos `include` y `exclude`. La sintaxis básica del método `select_dtypes()` es la siguiente:

    ```python
    DataFrame.select_dtypes(include=None, exclude=None)
    ```

    - `include` (opcional): Especifica los tipos de datos que se deben incluir en la selección. Puede ser una cadena de texto o una lista de tipos de datos, como `'int'`, `'float'`, `'object'`, etc. Si no se proporciona este argumento, se seleccionarán todas las columnas que coincidan con los tipos de datos especificados en `exclude`.

    - `exclude` (opcional): Especifica los tipos de datos que se deben excluir de la selección. Puede ser una cadena de texto o una lista de tipos de datos. Si no se proporciona este argumento, se seleccionarán todas las columnas que coincidan con los tipos de datos especificados en `include`.


    Algunos ejemplos de cómo se puede utilizar el método `select_dtypes()`:

    ```python
    # Seleccionar todas las columnas numéricas
    df.select_dtypes(include=['int', 'float'])

    # Seleccionar todas las columnas de tipo objeto (cadenas de texto)
    df.select_dtypes(include='object')

    # Seleccionar todas las columnas excepto las numéricas
    df.select_dtypes(exclude=['int', 'float'])

    # Seleccionar todas las columnas excepto las de tipo objeto (cadenas de texto)
    df.select_dtypes(exclude='object')
    ```

- Sacar los nombres de las columnas de tipo categórico. 

- Hacer un *for loop* para que nos haga los `.unique()` y los `.value_counts()` de cada columna

# sacamos las columnas que son de tipo object (categóricas) de nuestro DataFrame usando el método '.select_dtypes()'

df_cat = df.select_dtypes(include = "object")
df_cat.head()

Si miramos las columnas seleccionadas por Pandas nos damos cuenta que la mayoría tienen sentido, pero... que pasa con las columnas de 'cons.price.idx', 'cons.conf.idx', 'euribor3m', 'nr.employed' y 'date'. Si excluimos la columna de 'date', vemos claramente que son números, ¿qué esta pasando? Y es que, recordad que en Python los decimales se definen con ".", al ir estos números con comas automáticamente nos los convierte en *strings*. Esto ya aprenderemos a cambiarlo en próximas lecciones, pero ... ¿Cómo podríamos eliminar esas columnas para esta parte del análisis ya que no son columnas categóricas? Para eso usaremos el método `.drop()`. Este método se utiliza para eliminar filas o columnas de un DataFrame. Tiene la siguiente sintaxis básica:

```python
DataFrame.drop(labels, axis=0, inplace=False)
```

- `labels`: Especifica los nombres de las filas o columnas que se queremos eliminar. Puede ser un solo valor o una lista de valores.

- `axis`: Indica si se eliminarán filas (`axis=0`) o columnas (`axis=1`).

- `inplace` (opcional): Es un parámetro opcional que indica si se realizará la eliminación directamente en el DataFrame original (`inplace=True`) o si se creará un nuevo DataFrame con los cambios (`inplace=False`, valor por defecto).

En nuestro caso eliminaremos las columnas de 'ons.price.idx', 'cons.conf.idx', 'euribor3m', 'nr.employed' y 'date'. 


df_cat.head()

# eliminamos las columnas que no son de nuestro interés usando el método '.drop()'
# pasándolo una lista con las columnas que queremos eliminar
# axis = 1 porque los nombres que le hemos pasado en la lista corresponden a las columnas
# inplace = True porque queremos que los cambios realizados se sobreescriban en el DataFrame

df_cat.drop(['cons.price.idx', 'cons.conf.idx', 'euribor3m', 'nr.employed' , 'date'], axis = 1, inplace=True)

# chuequeamos como se ha quedado el DataFrame y vemos que ya lo tenemos preparado para poder evaluar cuales son sus valores únicos ('.unique()') y sus frecuencias ('.value_counts()')

df_cat.head()

# creamos una variable con los nombres de las columnas del DataFrame de las variables categóricas utilizando el método '.columns'

columnas_cat = df_cat.columns
print(f"Las columnas del DataFrame de variables categóricas son {columnas_cat}")

# empezamos a iterar por cada una de las columnas para sacar sus valores únicos y sus frecuencias

for columna in columnas_cat:
    print(f" \n----------- ESTAMOS ANALIZANDO LA COLUMNA: '{columna.upper()}' -----------\n")
    print(f"Sus valores únicos son: {df_cat[columna].unique()}\n")
    print(f"Las frecuencias de los valores únicos de las categorías son: {df_cat[columna].value_counts()} ")

Lo siguiente que deberíamos hacer es: 
- Tipos de variables que tenemos

- ¿Tenemos valores nulos?

- ¿Tenemos valores duplicados?

Manos a la obra!
