==> Mod3/jupyter_notebooks/03-02-EDA-valores-nulos.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd

# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# cargamos el csv con el que vamos a ir desarrollando la lección de hoy, recordad que la descripción la tenéis al inicio del jupyter. 
# para eso usaremos el método "pd.read_csv" que ya vimos en la leccion anterior.

df = pd.read_csv("files/bank-additional.csv", index_col = 0)

# hacemos un ".head()" para mostrar las 5 primeras filas del dataframe
df.head()

## Valores nulos


# ¿Cuántos valores nulos tenemos en la columna 'default' ?
# en este caso usaremos el método '.isnull()'
# como vemos nos devuelve una Serie que recordemos es una estructura de datos unidimensional que contiene datos etiquetados y se asemeja a una columna de un DataFrame

df["default"].isnull()

# lo que nos devuelve el método '.isnull()' por si mismo es poco entendible, 
# es por esto que nos suma todos los valores nulos que ha identificado como True usando el método '.sum()'
# en este caso tenemos 9016 valores nulos para la columna 'default'. 

df["default"].isnull().sum()

# en ejemplo anterior hemos visto el número de nulos para una sola columna,
# en este ejemplo lo haremos para todo el DataFrame, para que nos diga el número de nulos que hay por columna

df.isnull().sum()

En este ejemplo vemos que solo tenemos valores nulos para algunas columnas, en concreto para:

- age, la cual tiene 5120 nulos. 

- job, la cual tiene 345 nulos. 

- marital, la cual tiene 85 nulos. 

- education, la cual tiene 1807 nulos. 

- default, la cual tiene 8981 nulos.

- housing, la cual tiene 1026 nulos.

- loan, la cual tiene 1026 nulos, 

- cons.price.idx, la cual tienne 471 nulos. 

- euribor3m, la cual tiene 9256 nulos.

- date, la cual tiene 248 nulos. 


🚨 **NOTA** En Pandas existe el método `.isna()` que es exactamente igual que el `.isnull()`. Su sintaxis básica es:

```python
dataframe.isna()
# fijaos como es exactamente igual que el '.isnull()`
```

# veamos ahora el mismo ejemplo pero usando el metodo '.notnull()`
# que recordemos nos devuelve el número de valores no nulos que tenemos en cada columna
# en este caso, nos esta diciendo que en la columna "default" tenemos 34154 valores no nulos, es decir, 34019 con valores. 

df["default"].notnull().sum()

# si lo hacemos para todo el DataFrame

df.notnull().sum()

# hemos comentado que también podemos usar el método .info()' 
# que recordemos muestra información resumida sobre el DataFrame, incluyendo el número de datos, tipos de datos o valores no nulos

df.info()

Si interpretaramos los resultados obtenidos del método `.info()` que:

- Nos dice el número de filas que tenemos en el DataFrame, en este caso 43000

- Nos dice el número total de columnas que tenemos en el DataFrame, en este caso 23

- Los nombres de todas las columnas del DataFrame, en este caso 'age', 'job', 'marital', 'education', 'default', 'housing', 'loan', 'contact', 'duration', 'campaign', 'pdays', 'previous', 'poutcome', 'emp.var.rate', 'cons.price.idx', 'cons.conf.idx', 'euribor3m', 'nr.employed', 'y', 'month_day_week'

- El tipo de datos que tenemos en nuestro DataFrame, con el número de columnas que tenemos de cada tipo, en este caso float64(7), int64(4), object(12)

- Para cada una de las columnas tenemos:

    - El número de valores no nulos. En el caso de la columna `job` tiene 42655  valores no nulos

    - El tipo de la columna. En el caso de la columna `job` es de tipo object, que recordemos que es como el *string* con el que estamos familiarizadas.
