==> Mod3/jupyter_notebooks/03-03-EDA-valores-duplicados.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd

# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# cargamos el csv con el que vamos a ir desarrollando la lección de hoy, recordad que la descripción la tenéis al inicio del jupyter. 
# para eso usaremos el método "pd.read_csv" que ya vimos en la leccion anterior.

df = pd.read_csv("files/bank-additional.csv", index_col = 0)

# hacemos un ".head()" para mostrar las 5 primeras filas del dataframe
df.head()

## Valores duplicados



# recordemos el DataFrame, mostrando dos filas aleatorias

df.sample(2)

# ¿tenemos duplicados teniendo en cuenta todas las filas del DataFrame?
# al igual que en el caso del '.isnull()' nos devuelve una serie de Booleanos

df.duplicated()

# como hemos visto lo anterior nos devuelve una serie de booleanos lo cual es poco entendible, 
# por lo que añadiremos el método '.sum()' para que nos cuente cuantas filas tenemos duplicadas
# en este caso el código nos esta diciendo que no tenemos filas duplicadas

df.duplicated().sum()

# ¿Y si quisieramos saber si tenemos duplicados para una columna en concreto? 
# en ese caso tendremos que usar el parámetro subset()' 
# imaginemos que queremos saber si hay duplicados en la columna de 'date' 
# en este caso tenemos un total de 41139 fechas duplicadas en nuestro conjunto de datos (recordemos que teníamos 27570 datos)

df.duplicated(subset = "date").sum()

# si por ejemplo evaluaramos si existen valores duplicados para la columna de 'id_' debería salir que no existe ningún valor duplicado, ya que estos id deberían ser únicos
# usemos de nuevo el parámetro 'subset'
df.duplicated(subset  = "id_").sum()
