==> Mod3/jupyter_notebooks/04-01-metodos.ipynb <==
# Métodos de unión de datos

Como ya hemos comentado en Pandas tenemos tres métodos para unir datos,CONCAT, MERGE y JOIN, veamos como se comportan.


# importamos las librerías que necesitamos

# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import numpy as np


# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames


# antes de empezar vamos a leer los datos, el primero que importaremos será el csv donde tenemos la información de la campaña de marketing
df_mar = pd.read_csv("bank-additional.csv", index_col=0)
df_mar.head(1)

# vamos ahora a cargar el archivo de excel, recordemos que teníamos el método 'pd.read_excel()'
# además recuerda que este método, que vimos en la lección de este módulo, tiene un parámetro que se llama 'sheet_name' que nos permite elegir la hoja de trabajo que queremos abrir en Pandas
# 🚨 Nuestras hojas de trabajo se llaman '2012', '2013', '2014' 

df_2012 = pd.read_excel("customer-details.xlsx", sheet_name="2012", index_col = 0)
df_2013 = pd.read_excel("customer-details.xlsx", sheet_name="2013", index_col = 0)
df_2014 = pd.read_excel("customer-details.xlsx", sheet_name="2014", index_col = 0)

# comprobamos que los datos se han cargado correctamente
df_2012.head(1)

# miramos los datos del DataFrame df_2013
df_2013.head(1)

# chequeamos el último DataFrame
df_2014.head(1)

**¿Qué es lo que tenemos en todos los DataFrames que hemos cargado?**

- En el excel tenemos tres set de datos diferentes, pero que si nos fijamos tienen la misma información y columnas, con la diferencia de que en cada DataFrame tenemos los clientes que pasaron a ser clientes en 3 diferentes años. Estos los podríamos unificar en solo uno. 

- Además, estos conjuntos de datos tienen una columna con el 'id' personal de cada cliente, que si nos fijamos es lo mismo que la columna 'id_' del csv. En este caso podremos unirlos como hacíamos con los joins en SQL. 

## `concat()`

El método `concat()` puede ser utilizado para concatenar DataFrames o Series de pandas. Su sintaxis es:

```python
pd.concat(objs, axis=0, join='outer', ignore_index=False)
```

Para poner en práctica este tipo de unión usaremos los tres DataFrames que hemos cargado desde el Excel. El objetivo será tener toda la información personal de todos los clientes en un único DataFrame.

# recordamos los 3 DataFrames
df_2012.head(2)

df_2013.head(2)

df_2014.head(2)

# en este primer ejemplo vamos a unir los 3 DataFrames por las filas ('axis=0'). 
# Fijaos como al método 'pd.concat' le hemos pasado una lista con los nombres de los DataFrames que queremos unir
# usando el parámetro 'ignore_index = True' le decimos que reestablezca los índices del DataFrame nuevo generado, de esta forma los índices irán del 1-n (siendo n el número total de filas que tenemos en el nuevo DataFrame)
df_concat0 = pd.concat([df_2012, df_2013, df_2014], axis=0, ignore_index = True)
df_concat0.head()

Si nos fijamos, el número de filas del *DataFrame* concatenado `df_concat0` es el resultado de la suma de los tres *DataFrames*.

print(f"las filas y columnas que tenemos en el df_2012 son {df_2012.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_2013 son {df_2013.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_2014 son {df_2014.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_concat0 {df_concat0.shape}")
print("-------------------------------------------------------------------")

# pongamos ahora un ejemplo de concatenación cambiando del valor del parámetro axis. 
# si te fijas lo que hemos hecho es juntar los tres dataframes, uno al lado del otro.
df_concat1 = pd.concat([df_2012, df_2013, df_2014], axis=1)
df_concat1.head()

Como hemos cambiado el valor de axis, ahora la concatenacion se hace de manera horizontal y no vertical (apilado) como cuando habiamos establecido el axis = 0.

## `merge()` 


En este caso uniremos el DataFrame creado en el primer ejemplo del concat que vimos previamente y el DataFrame de la campaña de marketing, el que cargamos desde el *csv*

# recordemos el DataFrame de la campaña de marketing
df_mar.head(2)

# recordemosel DataFrame creado en el concat
df_concat0.head(2)

print(f"las filas y columnas que tenemos en el df concatenado son {df_concat0.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_mar son {df_mar.shape}")
print("-------------------------------------------------------------------")

Por otro lado, veamos los nombres de columnas que tenemos en los *DataFrames*.

print(f"los nombres de columnas que tenemos en el df concatenado son \n\n{df_concat0.columns}")
print("-------------------------------------------------------------------")


print(f"los nombres de columnas que tenemos en el df_mar son \n\n {df_mar.columns}")
print("-------------------------------------------------------------------")

Si nos fijamos, tenemos varias columnas diferentes entre los *DataFrames*, pero atencion! hay una columna que comparten! La de `id_` en el DataFrame de la campaña de marketing y la columna de 'ID' en el DataFrame concatenado. Esta columna será la que nos permita hacer el mergeo entre los dos *DataFrames*.

### Inner merge 

Cuando se realiza un *inner merge*, también conocido como intersección, se combinan los datos de dos *DataFrames* en función de una o más columnas comunes, pero solo se conservan las filas donde las columnas coinciden en ambos DataFrames. Como tenemos una columna en común pero que se llaman diferente en los *DataFrames* usaremos el parámetro `right_on` y `left_on`.

mergeado_inner = df_concat0.merge(df_mar, left_on='ID', right_on = "id_")
mergeado_inner.head()

print(f"las filas y columnas que tenemos en el df concatenado son {df_concat0.shape}")
print("-------------------------------------------------------------------")

print(f"las filas y columnas que tenemos en el df de marketing son {df_mar.shape}")
print("-------------------------------------------------------------------")

print(f"las filas y columnas que tenemos en el mergeado_inner son {mergeado_inner.shape}")
print("-------------------------------------------------------------------")



Fijaos que ahora tenemos nuestros *datasets* unidos, se nos han juntado las columnas de los *DataFrames* en aquellos valores que coinciden los valores de `id_`, que es por donde hemos especificado que mergee. 

> Tenemos el mismo número filas en el *DataFrame* mergeado que en el de la campaña de marketing.  Esto es así porque hay 'id'  que están en `df_concat0` y no en `df_mar`. Como no encuentra el "match" los elimina. Es decir no se ha llamado a todos los clientes. 
>
>El *DataFrame* `mergeado_inner` resultante contiene todas las columnas de `df_concat` así como todas las columnas de `df_mar`.
>
>También observamos que la columna del 'id_' y 'ID', es decir las columnas por las que hemos mergeado están incluidas en *DataFrame* final, es decir, tenemos columnas redundantes. 

```python
# 📌en caso de que las columnas que vamos a unir se hubieran llamado igual tendríamos que hacer usado la siguiente sintaxis, usando el parámetro 'on', en vez de 'right_on' y 'left_on': 

mergeado_inner2 = df_2002.merge(df_species, on = "NOMBRE_COLUMNA_COMÚN")
mergeado_inner2.head()
```

### Left merge

Cuando se realiza un *left merge*, también conocido como unión izquierda, se combinan los datos de dos *DataFrames* en función de una o más columnas comunes, y se conservan todas las filas del *DataFrame* de la izquierda (el primer *DataFrame* especificado en la fusión), incluyendo las filas que no tienen coincidencias en el *DataFrame* de la derecha.

mergeado_izq = df_concat0.merge( right=df_mar, how='left',
                        left_on='ID', right_on='id_')

mergeado_izq.head()

print(f"las filas y columnas que tenemos en el df_concat son {df_concat0.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el mergeado_izquierda son {mergeado_izq.shape}")
print("-------------------------------------------------------------------")

El *DataFrame* resultante de una *left join* se parece mucho al resultado del *inner join* en cuanto a las columnas que contiene. Sin embargo, a diferencia del *inner_join* (`mergeado_inner`), *left_join* (`mergeado_izq`) contiene el mismo número de filas que el *DataFrame* original de `df_concat0`. 

Cuando examinamos `mergeado_izq`, encontramos que hay filas en las que falta la información que debería proceder de `df_mar`. Es decir, se han generado 170 nulos en algunas columnas

> Tenemos 170 NaN , esto significa que hay 170 clientes que teníamos en el `df_concat0` que NO aparecían en `df_mar`, es decir, los clientes a los que no se contacto en la campaña de marketing. 

## `join()`

Se utiliza para combinar los datos de dos DataFrames en función de los índices de las filas. Proporciona una forma conveniente de realizar uniones basadas en índices en lugar de columnas.

La sintaxis básica del método `join()` es la siguiente:

```python
resultado = df_left.join(df_right, how='tipo_de_join',  lsuffix='', rsuffix='')



# volvamos a nuestros dataframes originales y comprobemos primero sus formas.

print(f"las filas y columnas que tenemos en el df_concat0 son {df_concat0.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_mar son {df_mar.shape}")
print("-------------------------------------------------------------------")

Lo primero que vamos a hacer es cambiar el nombre de la columna `id_` por `ID`, para eso usaremos el método `rename()`. Este método se utiliza para cambiar los nombres de columnas, índices o ambos en un DataFrame. Proporciona la capacidad de renombrar las etiquetas de las filas y columnas de manera flexible y conveniente.

La sintaxis básica del método `rename()` es la siguiente:

```python
dataframe.rename(columns=nombre_columnas, index=nombre_filas, inplace=False)
```

Donde:
- `dataframe`: Es el DataFrame al que deseas aplicar la operación de renombrar.

- `columns`: Es un diccionario o mapeo que especifica los nombres de columnas a cambiar. Las claves del diccionario representan los nombres actuales de las columnas, y los valores correspondientes son los nombres nuevos que deseas asignarles.

- `index` (opcional): Es un diccionario o mapeo que especifica los nombres de filas a cambiar. Las claves del diccionario representan los nombres actuales de las filas, y los valores correspondientes son los nombres nuevos que deseas asignarles.

- `inplace` (opcional): Es un parámetro booleano opcional que indica si se debe modificar el DataFrame original o si se debe crear un nuevo DataFrame con los nombres cambiados. Por defecto, su valor es `False`, lo que significa que se creará un nuevo DataFrame con los nombres modificados.

df_mar.head(1)

# cambiamos el nombre de la columna 'id_' por 'ID' en el DataFrame de df_mar usando el método rename()
df_mar.rename(columns = {"id_": "ID"}, inplace = True )

# chequeamos que el cambio se ha realizado
df_mar.head(2)


Una vez que hemos cambiado el nombre de una de las columnas lo siguiente que tenemos que hacer es establecer alguna de las columnas que tienen en común los dos *DataFrames* (`ID` o `id_`) como índice. Para eso vamos a usar un método que no habíamos visto todavía, el `set_index()`. Este método se utiliza para establecer una o más columnas como índices de un *DataFrame*. Cambia la estructura del DataFrame de modo que los valores de una o más columnas se conviertan en los índices de las filas.

La sintaxis básica del método `set_index()` es la siguiente:

```python
dataframe.set_index(keys, drop=True, inplace=False)
```

Donde:

- `dataframe`: Es el DataFrame al que deseas aplicar la función.

- `keys`: Especifica la columna o columnas que se utilizarán como índices. Puede ser un solo nombre de columna o una lista de nombres de columnas si se desean múltiples índices.

- `drop`(opcional): Es un parámetro booleano opcional que indica si se deben eliminar las columnas utilizadas como índices del DataFrame resultante. Por defecto, su valor es `True`, lo que significa que las columnas utilizadas como índices se eliminarán del DataFrame.

- `inplace` (opcional): Es un parámetro booleano opcional que indica si se debe modificar el DataFrame original o si se debe crear un nuevo DataFrame con los índices establecidos. Por defecto, su valor es `False`, lo que significa que se creará un nuevo DataFrame.

El método `set_index()` devuelve un nuevo DataFrame con los índices establecidos según las columnas especificadas. Si `inplace=True`, el método modifica el DataFrame original y no devuelve nada.

# utilizamos el método 'set_index()' para poner la columna de 'ID' como índice de nuestro DataFrame
df_mar.set_index(["ID"], inplace = True)

# vemos como ha quedado el DataFrame
df_mar.head(1)

# Veamos como estaba compuesto el dataframe df_concat0.
df_concat0.head()

# en este punto el orden es importante, y tendremos pasar primero el DataFrame que no tiene en el índice la columna y dentro del paréntesis el DataFrame que tiene la columna común en el índice. 
df_join = df_concat0.join(df_mar, on = "ID")
df_join.sample() #recuerden que sample es solo para traernos una muestra del dataframe

print(f"las filas y columnas que tenemos en el df_mar son {df_mar.shape}")
print("-------------------------------------------------------------------")


print(f"las filas y columnas que tenemos en el df_join son {df_join.shape}")
print("-------------------------------------------------------------------")
