==> Mod3/jupyter_notebooks/04-03-filtrado.ipynb <==
## Filtrado de datos


# nuestras librerias a utilizar
import pandas as pd
import numpy as np


# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# traemos los datos que guardamos en el tema anterior.

df_final = pd.read_csv("df_final_limpieza.csv")

# recordemos el DataFrame
df_final.head(2)

# imaginemos que queremos quedarnos solo con aquellos clientes que están casados 
# primero establecemos la condición de filtrado que queremos, en este caso que la columna de 'marital' sea igual a 'married'
condicion_casados = df_final["marital"] == "married"

# ahora aplicamos la condicion definida previamente a todo el DataFrame, esto lo haremos usando la sintaxis con corchetes
# vamos a almancenar los resultados en una variable que se llame 'df_casados'
df_casados = df_final[condicion_casados]

# comprobamos que el filtrado se hizo correctamente, en este caso, la columna 'marital' deberá tener un único valor  ('married')
df_casados["marital"].unique()

# visualizamos el DataFrame
df_casados.head()

# podríamos filtrar también con operadores de comparación, por ejemplo imaginemos que queremos quedarnos solo con los datos de aquellos clientes que tienen un 'income' mayor de 150000
condicion_income = df_final["income"] > 150000

# aplicamos el filtro al DataFrame
df_income_150 = df_final[condicion_income]

# comprobamos que el filtrado se hizo correctamente, en este caso el valor mínimo que tengamos para esta columna deberá ser mayor que 150000, para eso vamos a usar el método '.min()' de Pandas
df_income_150["income"].min()

df_income_150.head()

# ¿y si quisieramos juntar dos condiciones? Tendríamos que utilizar el operador "&" (si queremos un and) o el operador "|" (si queremos un or)
# en este caso vamos a buscar aquellos clientes que están casado Y que tienen un 'income' mayor que 150000

df_income_casados = df_final[condicion_casados & condicion_income]
df_income_casados.head(2)

### Método `.isin()`

df_final.head()

df_final["education"].unique()

# imaginemos que nos interesan solo los clientes que tienen un nivel de educación de 'basic 6y', 'basic 9y'. 
# podríamos usar lo métodos aprendidos previamente, pero la realidad es que cuantas más condiciones pongamos más probabilidad hay en equivocarnos. 
# para eso usaremos el método 'isin()' 

# primero definimos una lista con los valores que queremos filtrar
filtro_educacion = ['basic 6y', 'basic 9y']

# aplicamos el método 'isin()' pasando la lista
df_educacion = df_final[df_final["education"].isin(filtro_educacion)]

# comprobamos que ahora solo tenemos dos valores únicos para la columna ' education'
df_educacion["education"].unique()

### Método `.between()`



df_final.head(1)

# imaginemos que queremos saber que clientes han visitado nuestra página web entre 15 y 25 veces (inclusive), lo podríamos hacer con cualquiera de los métodos aprendidos hasta ahora, pero sería muy tedioso
# veamos como hacerlo con el método '.between()'

df_web = df_final[df_final["numwebvisitsmonth"].between(15, 25, inclusive = "both" )]
df_web.head()

# veamos como estan compuestas nuestras columnas, es decir de que tipo son sus valores
df_final.dtypes

# si nos fijamos la columna 'dt_customer' es de tipo object y nosotros nos guataria que sea del tipo datetime, por lo que usaremos el metodo to_datetime para especificar la columna a convertir.
df_final['dt_customer'] = pd.to_datetime(df_final['dt_customer'])

# nuevamente usamos dtypes sobre el dataframe para corroborar de que tipo es cada columna.
df_final.dtypes

# imaginemos que queremos saber que clientes entraron en el banco en enero del 2013. 
# lo primero que tendremos que hacer es generar dos objetos de tipo datetime con las fechas de inicio y fin de nuestro filtro
# para eso usaremos el método 'pd.to_datetime()' el cual necesita que le pasemos las fechas en formato string
inicio = pd.to_datetime('2013-01-01')
fin = pd.to_datetime('2013-01-31')

# después le pasaremos esas dos variables creadas al método '.between()' 
df_fechas = df_final[df_final["dt_customer"].between(inicio, fin, inclusive = "both")]
df_fechas.head(2)

### Método `.str.contains()`




df_final.head(2)

df_final.isnull().sum()

# imaginemos que queremos trabajar con la columna de 'education' y quisieramos sacar todos aquellos clientes que en su educación aparece un dígito
# lo primero que debemos hacer es sacar el patrón de regex
patron_regex = "\d+"

df_numeros = df_final[df_final["education"].str.contains(patron_regex, regex = True, na = False)]
df_numeros.head()

# para entender que hace bien el parámetro 'na', ejecutemos de nuevo el código cambiando el valor de 'na = True'
# en este caso vemos que tenemos algunas filas más, como la que tiene índice 7 que no aparecía en el ejercicio anterior. 
df_numeros = df_final[df_final["education"].str.contains(patron_regex, regex = True, na = True)]
df_numeros.head()

# antes de terminar la lección vamos a guardar este DataFrame que hemos hemos creado a lo largo de la lección para usarlo en la lección de mañana
# para eso usaremos el método 'pd.to_csv()' 

df_final.to_csv("bank-additional_full.csv",index=False)
