==> Mod3/jupyter_notebooks/05-01-apply.ipynb <==
# Apply



# nuestras librerias a utilizar
import pandas as pd
import numpy as np


# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# traemos los datos que guardamos en el tema anterior.

df = pd.read_csv("bank-additional_full.csv")

# Recordamos el dataframe con el que estamos trabajando

df.head()

## Categorizamos la columna de `age`

Vamos a categorizar la columna `age` para conseguir las siguientes categorías:


- Jóvenes adultos: 17 a 25 años

- Adultos jóvenes: 26 a 39 años

- Mediana edad: 40 a 59 años

- Adultos mayores: 60 en adelante


# lo primero que vamos a hacer es recordad cuál es valor máximo, mínimo de la columna de `age`
# para eso usaremos el método '.max()' y '.min()' de Pandas

print(f"El valor mínimo de la edad es: {df['age'].min()}")
print(f"El valor máximo de la edad es: {df['age'].max()}")

# lo primero que vamos a hacer es crear una función que reciba un número y nos devuelva a que grupo de edad pertenece

def categorizar_edad(numero):
    
    """
    Categoriza la edad en grupos específicos.

    Esta función toma como entrada un número que representa la edad y devuelve una categoría
    específica basada en el rango de edad.

    Args:
    ----------
    numero : int
        La edad que se desea categorizar.

    Returns:
    ------------
    str
        Una cadena que representa la categoría de edad a la que pertenece el número.

    """
    if numero >=17 and numero <= 25:
        return "Jóvenes adultos"
    
    elif numero >= 26 and numero <= 39:
        return "Adultos jóvenes"

    elif numero >= 40 and numero <= 59:
        return "Mediana edad"
    
    else:
        return "Adultos mayores"
    
# Probemos la función para ver si funciona correctamente:
print(f"Una persona de 56 pertenece a la categoría: {categorizar_edad(56)}")
print(f"Una persona de 20 pertenece a la categoría: {categorizar_edad(20)}")
print(f"Una persona de 33 pertenece a la categoría: {categorizar_edad(33)}")
print(f"Una persona de 45 pertenece a la categoría: {categorizar_edad(45)}")

# ya hemos creado la función y hemos visto que funciona, lo siguiente que tendríamos que hacer es aplicarla a todo nuestro DataFrame usando el método `apply()`
# si nos fijamos esto nos devuelve una Serie, pero este resultado no lo hemos almacenado en una variable, por lo que lo siguiente que haremos será crear una nueva columna en el DataFrame con el resultado de este apply
df["age"].apply(categorizar_edad)

# creamos una columna nueva que se llama "age_cat" resultado de nuestro 'apply()'
df["age_cat"] = df["age"].apply(categorizar_edad)

# chequeamos que se ha creado la columna nueva (siempre aparecerá al final del DataFrame)
df.head()

##  Convertir columnas de tipo *string* a *float*


- Lo primero que tenemos que tener en cuenta es que en Python los decimales van con "." y no con comas como tenemos en las columnas involucradas, por lo que tendremos que reemplazar las comas por puntos. 

- Una vez puestos los puntos, tendremos que convertir ese valor a float, para esto usaremos un método que no conocemos todavía que es el astype. 

# como antes, lo primero que hacemos es crear una función que reciba un string en el que tiene que cambiar las "," por "."

def cambiar_comas(cadena):
    """
    Convierte una cadena que representa un número decimal en formato internacional 
    (con comas como separador de miles y punto como separador decimal) a su 
    equivalente en formato de número flotante.

    Args:
        cadena (str): Una cadena que representa el número decimal en formato 
                      internacional.

    Returns:
        float: El número decimal equivalente en formato de número flotante.
    """
    return float(cadena.replace(",", "."))

# aplicamos la función a una de las columnas que nos interesa, por ejemplo 'conspriceidx'
df["conspriceidx"].apply(cambiar_comas)

# vaya ... nos ha devuelto error. El error nos dice que no puede aplicar el método replace a un float, y es que en estas columnas tenemos nulos, y los nulos para Pandas son de tipo float, 
# por lo que tendremos que modificar un poco la función para que nos funcione

def cambiar_comas(cadena):
    """
    Reemplaza las comas por puntos en una cadena dada que representa un número decimal
    en formato internacional (con comas como separador de miles y punto como separador decimal).

    Args:
        cadena (str): Una cadena que representa el número decimal en formato 
                      internacional.

    Returns:
        float: El número decimal equivalente en formato de número flotante.

    Note:
        Si ocurre algún error durante el proceso de reemplazo (por ejemplo, si el 
        argumento no es una cadena), la función devolverá np.nan (valor Not a Number) 
        para indicar un valor inválido o no disponible.
    """

    try:
        # Reemplazar las comas por puntos en la cadena
        return float(cadena.replace(",", "."))
    
    except:
        # Si ocurre algún error (por ejemplo, si el argumento no es una cadena),
        # devolver np.nan (valor Not a Number) para indicar un valor inválido o no disponible.
        return np.nan

# aplicamos de nuevo la función y sobreescribimos los resultados de nuestra función en la misma columna
df["conspriceidx"] = df["conspriceidx"].apply(cambiar_comas)

# ya tenemos las comas cambiadas para la columna 'conspriceidx', por lo que podríamos esperar que esta columna ya sea de tipo númerico. Comprobémoslo en la siguiente celda. 
df.head()

# chequeamos los tipos de datos de la columna para confirmar que ya es del tipo correcto. 
df["conspriceidx"].dtypes

# hemos hecho el cambio para una de las columnas, pero recordad que había dos columnas más que tenían este formato incorrecto
# vamos a crear una lista con los nombres de las columnas que queremos cambiar

lista_columnas = ['consconfidx', 'euribor3m']

# iteramos por la lista de columnas y a cada una de ellas le aplicamos la función que hemos creado con un '.apply()'
for col in lista_columnas:
    df[col] = df[col].apply(cambiar_comas)

# chequeamos que las columnas ya son del tipo correcto
df.dtypes

df.head()

# imagina ahora que quieres crear una columna nueva donde tengamos el resultado de sumar el número de hijos adolescentes ('teenhome') y niños ('kidhome'). En este caso la función tiene que recibir dos columnas, es decir, dos parámetros. 
# definir la función

def calcular_hijos_totales(col_niños, col_adolescentes):
    """
    Calcula el número total de hijos dados los números de niños y adolescentes.

    Parameters:
    col_niños (int): Número de niños.
    col_adolescentes (int): Número de adolescentes.

    Returns:
    str: Cadena que representa el número total de hijos junto con el texto "hijos".
    """
    hijos_totales = col_niños + col_adolescentes
    return str(hijos_totales) + " hijos"

# aplicar la función al DataFrame usando un apply y un lambda
df["hijos_totales"] = df.apply(lambda x: calcular_hijos_totales(x["kidhome"], x["teenhome"]), axis = 1)

# mostrar el DataFrame para ver si se creo correctamente la columna
df.head()

Entendamos el código. El parámetro `axis=1` indica que se aplicará la función a lo largo de las filas del DataFrame. Por otra parte, dentro de `apply`, se está utilizando una función lambda, la cual recibe un argumento `x`, que representa cada fila del DataFrame. En este caso, se llama a la función `calcular_hijos_totales` con los valores de las columnas "kidhome" y "teenhome" de cada fila específica (`x["kidhome"]` y `x["teenhome"]`).  Si esta sintaxis no la entiendes mucho, no te preocupes, mañana en clase la revisaremos en profundidad para entenderla mejor. 

# guardemos el codigo, puede que nos sirva mas adelante!

df.to_csv("bank-additional_full_apply.csv",index=False)
