==> Mod3/jupyter_notebooks/05-02-map-replace.ipynb <==
# Otros métodos de limpieza

- `.map()`:  Se utiliza para aplicar una transformación o reemplazo a cada elemento de una Serie, y devuelve una nueva Serie con los valores transformados o reemplazados.

- `.replace()`: Se utiliza para reemplazar valores en un DataFrame o Serie con otros valores especificados.


# pero antes....librerias!!!
# nuestras librerias a utilizar
import pandas as pd
import numpy as np


# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# y nuestros datos anteriores....
# traemos los datos que guardamos en el tema anterior.

df = pd.read_csv("bank-additional_full_apply.csv")

# lo primero que vamos a hacer es crear un diccionario donde:
# las claves sean los valores que tenemos en la columna sobre la que queremos aplicar el map
# los valores del diccionario serán los valores nuevos que queremos tener en la columna
# En nuestro caso queremos reemplazar los 0 por No y los 1 por Si

diccionario_mapa = {0: "No", 1: "Si"}

# una vez creado el diccionario, vamos a aplicar el map a la columna de 'loan'
# en este caso vamos a sobreescribir los valores de la columna

df["loan"] = df["loan"].map(diccionario_mapa)

# vemos el DataFrame y veremos que ahora la columna de 'loan' ya no tiene 1 y 0, ahora tiene Si y No. 
df.sample(6)

# chequeamos los valores únicos para la columna de loan 
df["loan"].unique()

# hacemos lo mismo para la columna de 'housing' y 'default', para la cual nos vale el mismo diccionario que hemos creado antes. 
df["housing"] = df["housing"].map(diccionario_mapa)
df["default"] = df["default"].map(diccionario_mapa)


# vemos cuáles son sus valores únicos de la columna 'housing'
df["housing"].unique()

# vemos cuáles son sus valores únicos de la columna 'housing'
df["default"].unique()

## `.replace()`


El método `replace()` se utiliza para reemplazar valores en un *DataFrame* o una *Serie* con otros valores específicos que proporcionamos. Puedes utilizar este método para realizar cambios en los datos de manera eficiente y fácil.

La sintaxis básica del `replace()` en Pandas es la siguiente:

```python
DataFrame.replace(to_replace=None, value=None, inplace=False, limit=None, regex=False, method='pad')
```


df.head(1)

# aplicamos el método '.replace()' sobre la columna 'pdays' para reemplazar los valores 999 por nulos de numpy (np.nan)
df["pdays"] = df["pdays"].replace(999, np.nan)

# comprobamos que no tenemos valores de 999 en la columna 'pdays' . Perfecto! No nos sale ninguno
df[df["pdays"] == 999]

# guardamos el csv para usarlo en la lección de mañana
df.to_csv("bank-additional_clean.csv")
