==> Mod3/jupyter_notebooks/08-05-medidas-relativas.ipynb <==
# importamos las librerías que necesitamos

# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import numpy as np

# Visualización
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Evaluar linealidad de las relaciones entre las variables
# ------------------------------------------------------------------------------
from scipy.stats import shapiro, kstest

# Configuración
# -----------------------------------------------------------------------
pd.set_option('display.max_columns', None) # para poder visualizar todas las columnas de los DataFrames

# Gestión de los warnings
# -----------------------------------------------------------------------
import warnings
warnings.filterwarnings("ignore")


# lo primero que hacemos es cargar el dataframe que vamos a usar para esta lección
df = pd.read_csv("../files/bank-additional-clean-nonulls.csv", index_col = 0)
df.head()

# Medidas de posición relativas

En este notebook, aprenderemos sobre las medidas de posición relativas. Estas medidas nos ayudan a comprender la distribución de los datos y a identificar valores atípicos.

## Percentiles

Para calcular los percentiles, usaremos el método `np.percentile()`.  La sintaxis de esta función es:

```python
np.percentile(a, q)
```

# Calculamos los percentiles 25 y 75 de la Columna 'age'
percentile_25_75 = np.percentile(df["age"], [25, 75])

# mostramos los resultados
print(f'El percentil 25 de la columna "age" es: {percentile_25_75[0]}')
print(f'El percentil 75 de la columna "age" es: {percentile_25_75[1]}')

# Interpretación
# El percentil 25 (primer cuartil) nos indica que el 25% de los clientes tienen una edad igual o menor que 33.
# El percentil 75 (tercer cuartil) nos indica que el 75% de clientes tienen una edad igual o menor a 46. 


df.head()
